+++ app/views/user_management/disability/edit.html.erb
@@ -6,7 +6,7 @@
         <strong><%= t('.form_title') %></strong>
       </div>
       <div class='card-body'>
-        <%= simple_form_for(@user, url: user_disability_path, html: { method: :put }) do |f| %>
+        <%= simple_form_for(@user, url: disability_user_path, html: { method: :put }) do |f| %>
           <div class='row'>
             <div class='form-group col-sm-12'>
               <%= f.error_notification %>
