+++ app/views/user_management/users/_disability.html.erb
@@ -2,7 +2,7 @@
   <div class='card-header'>
     <%= fa_icon 'user', text: t('.disability_status') %>
     <div class='card-header-actions'>
-      <%= link_to_edit(user_disability_path(@user)) %>
+      <%= link_to_edit(disability_user_path(@user)) %>
     </div>
   </div>
