+++ config/locales/controllers/user_management/disability.tr.yml
@@ -1,5 +1,5 @@
 tr:
-  account:
+  user_management:
     disability:
       edit:
         form_title: Engel Durumunu Güncelle
