+++ config/locales/models/user/en.yml
@@ -53,54 +53,55 @@ en:
     placeholders:
       user:
         <<: *user_attributes
-  users:
-    account:
-      <<: *user_attributes
-    create:
-      success: User successfully created.
-    destroy:
-      success: User successfully deleted!
-      warning: User can not be deleted.
-    disability:
-      disability_status: Disability Status
-      rate: Disability Rate
-    edit:
-      form_title: Edit Account
-    employees:
-      active: Is Active?
-      article: Article of Duty
-      duties: Duties
-      employees: Employee Records
-      positions: Positions
-      <<: *employee_attributes
-      <<: *duty_attributes
-      <<: *position_attributes
-      title: Title
-    index:
-      <<: *user_attributes
-      addresses: Addresses
-      card_header: User Accounts
-      first_name: First Name
-      identities: Identities
-      last_name: Last Name
-      new_user_link: Create a New User
-      smart_search_placeholder: TC/YU number, e-mail address, first of last name
-    new:
-      form_title: Create Account
-    save_address_from_mernis:
-        wait: The address was recently created or updated. You can update the address once a week.
-        will_update: We have received your request to create or update addresses, the address will be updated within a few seconds.
-    save_identity_from_mernis:
-        wait: The identity was recently created or updated. You can update the identity once a week.
-        will_update: We have received your request for an identity creation or an update, and within a few seconds/minutes the identity will be updated.
-    show:
-      account: Account
-      addresses: Addresses
-      employees: Employees
-      identities: Identities
-      other_information: Other Information
-      students: Student Information
-    students:
-      <<: *student_attributes
-    update:
-      success: User successfully updated.
+  user_management:
+    users:
+      account:
+        <<: *user_attributes
+      create:
+        success: User successfully created.
+      destroy:
+        success: User successfully deleted!
+        warning: User can not be deleted.
+      disability:
+        disability_status: Disability Status
+        rate: Disability Rate
+      edit:
+        form_title: Edit Account
+      employees:
+        active: Is Active?
+        article: Article of Duty
+        duties: Duties
+        employees: Employee Records
+        positions: Positions
+        <<: *employee_attributes
+        <<: *duty_attributes
+        <<: *position_attributes
+        title: Title
+      index:
+        <<: *user_attributes
+        addresses: Addresses
+        card_header: User Accounts
+        first_name: First Name
+        identities: Identities
+        last_name: Last Name
+        new_user_link: Create a New User
+        smart_search_placeholder: TC/YU number, e-mail address, first of last name
+      new:
+        form_title: Create Account
+      save_address_from_mernis:
+          wait: The address was recently created or updated. You can update the address once a week.
+          will_update: We have received your request to create or update addresses, the address will be updated within a few seconds.
+      save_identity_from_mernis:
+          wait: The identity was recently created or updated. You can update the identity once a week.
+          will_update: We have received your request for an identity creation or an update, and within a few seconds/minutes the identity will be updated.
+      show:
+        account: Account
+        addresses: Addresses
+        employees: Employees
+        identities: Identities
+        other_information: Other Information
+        students: Student Information
+      students:
+        <<: *student_attributes
+      update:
+        success: User successfully updated.
