+++ config/locales/models/user/tr.yml
@@ -53,67 +53,68 @@ tr:
     placeholders:
       user:
         <<: *user_attributes
-  users:
-    account:
-      <<: *user_attributes
-    addresses:
-      create_from_mernis: MERNİS'ten Oluştur
-      formal_address: Resmi Adresiniz
-      informal_address: Diğer Adresiniz
-      new_address: Yeni Adres Ekle
-      update_from_mernis: MERNİS'ten Güncelle
-    create:
-      success: Kullanıcı başarıyla oluşturuldu.
-    destroy:
-      success: Kullanıcı başarıyla silindi!
-      warning: Kullanıcı silinemedi!
-    disability:
-      disability_status: Engel Durumu
-      rate: Engel Oranı
-    edit:
-      form_title: Hesabı Düzenle
-    employees:
-      active: Aktif mi?
-      article: Görevlendirme Kanun Maddesi
-      duties: Görevlendirmeler
-      employees: Personel Kayıtları
-      positions: İdari Görevler
-      <<: *employee_attributes
-      <<: *duty_attributes
-      <<: *position_attributes
-      title: Unvan
-    identities:
-      create_from_mernis: MERNİS'ten Oluştur
-      informal_identity: Diğer Kimliğiniz
-      new_identity: Yeni Kimlik Ekle
-      update_from_mernis: MERNİS'ten Güncelle
-      student_identity: Öğrenci Kimliğiniz
-      user_identity: Kullanıcı Kimliğiniz
-    index:
-      <<: *user_attributes
-      addresses: Adres Bilgileri
-      card_header: Kullanıcı Hesapları
-      first_name: Adı
-      identities: Kimlik Bilgileri
-      last_name: Soyadı
-      new_user_link: Yeni Kullanıcı Oluştur
-      smart_search_placeholder: TC/YU numarası, e-posta adresi, ad veya soyad
-    new:
-      form_title: Hesap Oluştur
-    save_address_from_mernis:
-        wait: Adres yakın bir zamanda oluşturulmuş veya güncellenmiş. Adresi haftada bir güncelleyebilirsiniz.
-        will_update: Adres oluşturma veya güncelleme talebinizi aldık, bir kaç saniye/dakika içerisinde adres güncellenecek.
-    save_identity_from_mernis:
-        wait: Kimlik yakın bir zamanda oluşturulmuş veya güncellenmiş. Kimliği haftada bir güncelleyebilirsiniz.
-        will_update: Kimlik oluşturma veya güncelleme talebinizi aldık, bir kaç saniye/dakika içerisinde kimlik güncellenecek.
-    show:
-      account: Hesap Bilgileri
-      addresses: Adres Bilgileri
-      employees: Personel İşlemleri
-      identities: Kimlik Bilgileri
-      other_information: Diğer Bilgiler
-      students: Öğrenci Bilgileri
-    students:
-      <<: *student_attributes
-    update:
-      success: Kullanıcı başarıyla güncellendi.
+  user_management:
+    users:
+      account:
+        <<: *user_attributes
+      addresses:
+        create_from_mernis: MERNİS'ten Oluştur
+        formal_address: Resmi Adresiniz
+        informal_address: Diğer Adresiniz
+        new_address: Yeni Adres Ekle
+        update_from_mernis: MERNİS'ten Güncelle
+      create:
+        success: Kullanıcı başarıyla oluşturuldu.
+      destroy:
+        success: Kullanıcı başarıyla silindi!
+        warning: Kullanıcı silinemedi!
+      disability:
+        disability_status: Engel Durumu
+        rate: Engel Oranı
+      edit:
+        form_title: Hesabı Düzenle
+      employees:
+        active: Aktif mi?
+        article: Görevlendirme Kanun Maddesi
+        duties: Görevlendirmeler
+        employees: Personel Kayıtları
+        positions: İdari Görevler
+        <<: *employee_attributes
+        <<: *duty_attributes
+        <<: *position_attributes
+        title: Unvan
+      identities:
+        create_from_mernis: MERNİS'ten Oluştur
+        informal_identity: Diğer Kimliğiniz
+        new_identity: Yeni Kimlik Ekle
+        update_from_mernis: MERNİS'ten Güncelle
+        student_identity: Öğrenci Kimliğiniz
+        user_identity: Kullanıcı Kimliğiniz
+      index:
+        <<: *user_attributes
+        addresses: Adres Bilgileri
+        card_header: Kullanıcı Hesapları
+        first_name: Adı
+        identities: Kimlik Bilgileri
+        last_name: Soyadı
+        new_user_link: Yeni Kullanıcı Oluştur
+        smart_search_placeholder: TC/YU numarası, e-posta adresi, ad veya soyad
+      new:
+        form_title: Hesap Oluştur
+      save_address_from_mernis:
+          wait: Adres yakın bir zamanda oluşturulmuş veya güncellenmiş. Adresi haftada bir güncelleyebilirsiniz.
+          will_update: Adres oluşturma veya güncelleme talebinizi aldık, bir kaç saniye/dakika içerisinde adres güncellenecek.
+      save_identity_from_mernis:
+          wait: Kimlik yakın bir zamanda oluşturulmuş veya güncellenmiş. Kimliği haftada bir güncelleyebilirsiniz.
+          will_update: Kimlik oluşturma veya güncelleme talebinizi aldık, bir kaç saniye/dakika içerisinde kimlik güncellenecek.
+      show:
+        account: Hesap Bilgileri
+        addresses: Adres Bilgileri
+        employees: Personel İşlemleri
+        identities: Kimlik Bilgileri
+        other_information: Diğer Bilgiler
+        students: Öğrenci Bilgileri
+      students:
+        <<: *student_attributes
+      update:
+        success: Kullanıcı başarıyla güncellendi.
