+++ test/policies/user_management/disability_policy_test.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'pundit_test_case'
+
+module UserManagement
+  class DisabilityPolicyTest < PunditTestCase
+    %w[
+      edit?
+      update?
+    ].each do |method|
+      test method do
+        assert_permit     users(:serhat)
+        assert_not_permit users(:mine)
+      end
+    end
+  end
+end
